requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
apscheduler=3.6.3=pypi_0
arviz=0.6.1=pypi_0
attrs=19.3.0=pypi_0
autoimpute=0.11.6=pypi_0
beautifulsoup4=4.9.1=pypi_0
bs4=0.0.1=pypi_0
cachetools=4.2.4=pypi_0
cassandra-driver=3.25.0=pypi_0
certifi=2019.11.28=pypi_0
cffi=1.15.0=pypi_0
cftime=1.0.4.2=pypi_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
colorhash=1.0.2=pypi_0
configparser=4.0.2=pypi_0
cycler=0.10.0=pypi_0
flask=1.1.1=pypi_0
flask-cors=3.0.8=pypi_0
flask-monitoringdashboard=3.0.6=pypi_0
geomet=0.2.1.post1=pypi_0
gevent=21.8.0=pypi_0
google-api-core=2.2.1=pypi_0
google-api-python-client=2.28.0=pypi_0
google-auth=2.3.2=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-cloud-core=2.1.0=pypi_0
google-cloud-storage=1.42.3=pypi_0
google-crc32c=1.3.0=pypi_0
google-resumable-media=2.1.0=pypi_0
googleapis-common-protos=1.53.0=pypi_0
greenlet=1.1.2=pypi_0
gunicorn=20.0.4=pypi_0
h5py=2.10.0=pypi_0
httplib2=0.20.1=pypi_0
idna=2.10=pypi_0
imbalanced-learn=0.8.1=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=1.4.0=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.11.0=pypi_0
joblib=0.14.1=pypi_0
jsonschema=3.2.0=pypi_0
kiwisolver=1.1.0=pypi_0
kneed=0.5.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.2=pypi_0
missingno=0.4.2=pypi_0
more-itertools=8.1.0=pypi_0
netcdf4=1.5.3=pypi_0
numpy=1.18.1=pypi_0
oauthlib=3.1.1=pypi_0
packaging=20.1=pypi_0
pandas=0.25.3=pypi_0
patsy=0.5.1=pypi_0
pillow=8.3.2=pypi_0
pip=21.0.1=py36haa95532_0
protobuf=3.19.1=pypi_0
psutil=5.6.7=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pymc3=3.8=pypi_0
pymongo=3.11.3=pypi_0
pyparsing=2.4.6=pypi_0
pyrsistent=0.15.7=pypi_0
python=3.6.12=h5500b2f_2
python-dateutil=2.8.1=pypi_0
pytz=2019.3=pypi_0
pyyaml=5.3=pypi_0
regexp=0.1=pypi_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.4.1=pypi_0
seaborn=0.10.0=pypi_0
setuptools=52.0.0=py36haa95532_0
six=1.14.0=pypi_0
sklearn=0.0=pypi_0
soupsieve=2.0.1=pypi_0
sqlalchemy=1.3.13=pypi_0
sqlite=3.33.0=h2a8f88b_0
statsmodels=0.11.0=pypi_0
theano=1.0.4=pypi_0
threadpoolctl=3.0.0=pypi_0
tqdm=4.42.1=pypi_0
tzlocal=2.0.0=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=1.25.10=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=0.16.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
xarray=0.15.0=pypi_0
xgboost=0.90=pypi_0
zipp=2.0.1=pypi_0
zlib=1.2.11=h62dcd97_4
zope-event=4.5.0=pypi_0
zope-interface=5.4.0=pypi_0
